Properly implemented CreateFolder
