--- client/components/Application/Dialogs/AddEmptyFolder.tsx
@@ -1,21 +1,0 @@
-import InputDialog from '../../Parts/Dialogs/Input'
-import * as store from '@client/store'
-
-export default function AddEmptyFolderDialog() {
-  const folderPath = store.useStore(store.getFolderPath)
-
-  return (
-    <InputDialog
-      open={true}
-      value={folderPath}
-      title="Create new folder"
-      label="Create"
-      placholder="Name of the new folder"
-      onCancel={store.closeDialog}
-      onConfirm={async (path) => {
-        await store.createFolder(path)
-        store.closeDialog()
-      }}
-    />
-  )
-}
