+++ client/components/Application/Dialogs/CreateFolder/index.ts
@@ -1,0 +1,1 @@
+export { CreateFolderDialog } from './CreateFolder'
