--- client/store/actions/folder.ts
@@ -1,15 +1,0 @@
-import { client } from '@client/client'
-import * as store from '../store'
-import { onFileCreated } from './file'
-
-export async function createFolder(path: string) {
-  const result = await client.folderCreate({ path, deduplicate: true })
-
-  if (result instanceof client.Error) {
-    return store.setState('create-folder-error', (state) => {
-      state.error = result
-    })
-  }
-
-  await onFileCreated([result.path])
-}
