+++ client/store/actions/index.ts
@@ -4,7 +4,6 @@ export * from './dialog'
 export * from './error'
 export * from './event'
 export * from './file'
-export * from './folder'
 export * from './panel'
 export * from './project'
 export * from './refs'
